Exception handling for writer.
The Palworld 'backup' folder inside the game save folder location ---after patch 0.1.4.1--- is changing a lot. It creates an exception that creates .NET window. It's harmless but annoying. The errors while copying data now are creating the error file. The separate label (label5) will inform user about it. The label gets cleaned after start of the manual or automatic backup.